Split cli.py into separate function classes within daeclipsecli/
