+++ daeclipsecli/__init__.py
@@ -1,0 +1,7 @@
+"""daeclipsecli __init__.py."""
+
+from .add_art_to_groups import add_art_to_groups
+from .hot_tags import hot_tags
+from .post_status import post_status
+from .show_tags import show_tags
+from .user_comments import user_comments
